move helical simulator to downstream package
