+++ src/cryojax/simulator/_structural_ensemble/__init__.py
@@ -1,9 +1,6 @@
 from .assembly import (
     AbstractAssembly as AbstractAssembly,
     AbstractAssemblyWithSubunit as AbstractAssemblyWithSubunit,
-    compute_helical_lattice_positions as compute_helical_lattice_positions,
-    compute_helical_lattice_rotations as compute_helical_lattice_rotations,
-    HelicalAssembly as HelicalAssembly,
 )
 from .base_conformation import (
     AbstractConformationalVariable as AbstractConformationalVariable,
