+++ src/cryojax/simulator/_structural_ensemble/assembly/__init__.py
@@ -2,8 +2,3 @@ from .base_assembly import (
     AbstractAssembly as AbstractAssembly,
     AbstractAssemblyWithSubunit as AbstractAssemblyWithSubunit,
 )
-from .helix import (
-    compute_helical_lattice_positions as compute_helical_lattice_positions,
-    compute_helical_lattice_rotations as compute_helical_lattice_rotations,
-    HelicalAssembly as HelicalAssembly,
-)
